add -o options

+++ panda/src/downloadertools/pdecrypt.cxx
@@ -53,8 +53,10 @@ do_decrypt(istream &read_stream, ostream &write_stream) {
 void 
 usage() {
   cerr
+    << "\nUsage:\n"
+    << "   pdecrypt file.pe [file2.pe file3.pe ...]\n"
+    << "   pdecrypt -o dest_file file.pe\n\n"
     << "\n"
-    << "Usage: pdecrypt [opts] file [file2 file3 ...]\n\n"
     
     << "This program reverses the operation of a previous pencrypt command.  It\n"
     << "decrypts the contents of the named source file(s) and removes the .pe\n"
@@ -75,12 +77,20 @@ int
 main(int argc, char *argv[]) {
   extern char *optarg;
   extern int optind;
-  const char *optstr = "p:h";
+  const char *optstr = "o:p:h";
+
+  Filename dest_filename;
+  bool got_dest_filename = false;
 
   int flag = getopt(argc, argv, optstr);
 
   while (flag != EOF) {
     switch (flag) {
+    case 'o':
+      dest_filename = Filename::from_os_specific(optarg);
+      got_dest_filename = true;
+      break;
+
     case 'p':
       password = optarg;
       got_password = true;
@@ -103,16 +113,24 @@ main(int argc, char *argv[]) {
     return 1;
   }
 
+  if (got_dest_filename && argc > 2) {
+    cerr << "Only one input file allowed in conjunction with -o.\n";
+    return 1;
+  }
+
   bool all_ok = true;
   for (int i = 1; i < argc; i++) {
     Filename source_file = Filename::from_os_specific(argv[i]);
-    if (source_file.get_extension() != "pe") {
+    if (!got_dest_filename && source_file.get_extension() != "pe") {
       cerr << source_file 
            << " doesn't end in .pe; can't derive filename of output file.\n";
       all_ok = false;
 
     } else {
-      Filename dest_file = source_file.get_fullpath_wo_extension();
+      Filename dest_file = dest_filename;
+      if (!got_dest_filename) {
+        dest_file = source_file.get_fullpath_wo_extension();
+      }
 
       // Open source file
       ifstream read_stream;
@@ -149,7 +167,9 @@ main(int argc, char *argv[]) {
             dest_file.unlink();
             
           } else {
-            source_file.unlink();
+            if (!got_dest_filename) {
+              source_file.unlink();
+            }
           }
         }
       }
